Fix broken links in documentation
